Rearranged assembler files, and tried to fix jump handling
