+++ test_programs/hello_world.tri
@@ -1,1 +1,1 @@
-mbm AhK DBm mbm fJM DBm mbm faj DBm mbm HhJ DBm mbm fdA DBm mbm fKf DBm mbm Hhi DBm mbm MMM DBm 000
+K0m AhK aAm K0m fJM aAm K0m faj aAm K0m HhJ aAm K0m fdA aAm K0m fKf aAm K0m Hhi aAm K0m MMM aAm 000
