+++ test_programs/loop_test.tas
@@ -1,7 +1,14 @@
+func1:
+  SHOW A
+  SHOW B
+end func1
+
 main:
 
-  !loop
-  STRPNT "LOOK AROUND YOU"
-  JP loop
+  SET A, 3
+  SET B, 9
+  JP test
+  !test
+  CALL func1
 
 end main
