+++ test_programs/loop_test.tri
@@ -1,1 +1,1 @@
-K0m 0AG aAm K0m 0mE aAm K0m HfJ aAm K0m a0e aAm K0m bMj aAm K0m Bmb aAm K0m bfj aAm K0m akm aAm 0jj 000 000
+k0D 000 000 00c k0C 000 000 00i 0jj 00j 0jm 0aM 000 aaD aaC 0jM
