+++ triangulate/test.py
@@ -1,0 +1,12 @@
+import triangulate
+import handle_instr
+
+def test_add_trytes():
+    output = triangulate.triangulate("main:\nADD A0, B0\nend main", False)
+    assert(output == "AMJ 000")
+    output = triangulate.triangulate("main:\nADD E2, C1\nend main", False)
+    assert(output == "AaF 000")
+
+def test_set_trytes():
+    output = triangulate.triangulate("main:\nSET A0, 7\nend main", False)
+    assert(output == "K0M 00g 000")
